+++ opps/contrib/logging/admin.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+from django.contrib import admin
+
+from .models import Logging
+
+
+class LoggingAdmin(admin.ModelAdmin):
+    model = Logging
+    raw_id_fields = ('user',)
+    exclude = ('site_iid', 'site_domain')
+
+
+admin.site.register(Logging, LoggingAdmin)