+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Changelog
 * Fix image_obj template tag when sending Nonetype image
 * Create Opps Vagrant box to help other contributors
 * Added support to ajax requests with extends_parent variable in template context
+* Write logging contrib application #275
 
 0.2.2
 =====